restructuring to make better use of blueprints (added a core module)
