+++ monopyly/auth/routes.py
@@ -1,24 +1,15 @@
 """
-Flask blueprint for site authentication.
+Routes for site authentication.
 """
-import functools
-from flask import (
-    Blueprint, flash, g, redirect, render_template, request, session, url_for
-)
+from flask import flash, redirect, render_template, request, session, url_for
 from werkzeug.security import check_password_hash, generate_password_hash
 
-from .db import get_db
+from monopyly.db import get_db
+from monopyly.auth.tools import get_username_and_password
+from monopyly.auth import auth
 
 
-bp = Blueprint('auth', __name__, url_prefix='/auth')
-
-def get_username_and_password(form):
-    """Get username and password from a form."""
-    username = form['username']
-    password = form['password']
-    return username, password
-
-@bp.route('/register', methods=('GET', 'POST'))
+@auth.route('/register', methods=('GET', 'POST'))
 def register():
     if request.method == 'POST':
         # Get username and passwords from the form
@@ -50,7 +41,7 @@ def register():
     return render_template('auth/register.html')
 
 
-@bp.route('/login', methods=('GET', 'POST'))
+@auth.route('/login', methods=('GET', 'POST'))
 def login():
     if request.method == 'POST':
         # Get username and passwords from the form
@@ -77,28 +68,8 @@ def login():
     # Display the login page
     return render_template('auth/login.html')
 
-@bp.route('/logout')
+@auth.route('/logout')
 def logout():
     # End the session and clear the user ID;k
     session.clear()
     return redirect(url_for('index'))
-
-@bp.before_app_request
-def load_logged_in_user():
-    # Match the user's information with the session
-    user_id = session.get('user_id')
-    if user_id is None:
-        g.user = None
-    else:
-        user_query = 'SELECT * FROM users WHERE id = ?'
-        db = get_db()
-        cursor = db.cursor()
-        g.user = cursor.execute(user_query, (user_id,)).fetchone()
-
-def login_required(view):
-    @functools.wraps(view)
-    def wrapped_view(**kwargs):
-        if g.user is None:
-            return redirect(url_for('auth.login'))
-        return view(**kwargs)
-    return wrapped_view
