+++ monopyly/core/__init__.py
@@ -1,0 +1,10 @@
+"""
+Flask blueprint for core functionality.
+"""
+from flask import Blueprint
+
+# Define the blueprint
+core = Blueprint('core', __name__)
+
+import monopyly.core.routes
+import monopyly.core.filters
