+++ monopyly/templates/layout.html
@@ -39,15 +39,15 @@
 		<header id="masthead">
 			<div class="container">
 
-				<a href="{{ url_for('index') }}">
+				<a href="{{ url_for('core.index') }}">
 					<h1 class="monopyly-logo">Monopyly</h1>
 				</a>
 
 				<nav id="nav-menu">
 					<ul></strong>
 
-						<li><a href="{{ url_for('index') }}">Home</a></li>
-						<li><a href="{{ url_for('about') }}">About</a></li>
+						<li><a href="{{ url_for('core.index') }}">Home</a></li>
+						<li><a href="{{ url_for('core.about') }}">About</a></li>
 						{% if g.user %}
 		      		<li><span>{{ g.user['username'] }}</span>
 		      		<li><a href="{{ url_for('auth.logout') }}">Log Out</a>
